renamed : directory name to Eloquents from Eloquent.
