+++ app/Infrastructure/Eloquents/BookAnchors.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Infrastructure\Eloquent;
+namespace App\Infrastructure\Eloquents;
 
 use Illuminate\Database\Eloquent\Model;
 use Alsofronie\Uuid\Uuid32ModelTrait;
 
-class Tag extends Model
+class BookAnchors extends Model
 {
     use Uuid32ModelTrait;
